+++ hpOneView/resources/resource.py
@@ -1209,7 +1209,7 @@ class ResourceClient(object):
 
         return self.__do_post(uri, resource, timeout, custom_headers)
 
-    def upload(self, file_path, uri=None, timeout=-1):
+    def upload(self, file_path, uri=None, custom_headers=None, timeout=-1):
         """
         Makes a multipart request.
 
@@ -1229,7 +1229,7 @@ class ResourceClient(object):
             uri = self._uri
 
         upload_file_name = os.path.basename(file_path)
-        task, entity = self._connection.post_multipart_with_response_handling(uri, file_path, upload_file_name)
+        task, entity = self._connection.post_multipart_with_response_handling(uri, file_path, custom_headers, upload_file_name)
 
         if not task:
             return entity