+++ hpOneView/resources/settings/firmware_bundles.py
@@ -45,7 +45,7 @@ class FirmwareBundles(object):
         self._connection = con
         self._client = ResourceClient(con, self.URI)
 
-    def upload(self, file_path, timeout=-1):
+    def upload(self, file_path, custom_headers=None, timeout=-1):
         """
         Upload an SPP ISO image file or a hotfix file to the appliance.
         The API supports upload of one hotfix at a time into the system.
@@ -59,4 +59,5 @@ class FirmwareBundles(object):
         Returns:
           dict: Information about the updated firmware bundle.
         """
-        return self._client.upload(file_path, timeout=timeout)
+#        custom_headers = { 'initialScopeUris': '/rest/scopes/bf3e77e3-3248-41b3-aaee-5d83b6ac4b49'}
+        return self._client.upload(file_path, custom_headers=custom_headers, timeout=timeout)